refactor: tracklist
