+++ src/entity/songs/index.ts
@@ -1,5 +1,3 @@
-export * from "./ui";
-
 export * as songModel from "./model";
 
 export * from "./model";
