--- src/entity/songs/ui/index.ts
@@ -1,1 +1,0 @@
-export * from "./tracklist-item";
