+++ src/features/add-to-playlist/view.tsx
@@ -1,9 +1,11 @@
-import { PlusSmIcon } from "@heroicons/react/outline";
 import { useUnit } from "effector-react";
+import { memo } from "react";
 
-import type { Song } from "@/src/entity/songs/lib";
+import type { Song } from "@/src/entity/songs";
 
 import { addToPlaylistButtonClicked } from "./model";
+
+import { PlusSmIcon } from "@heroicons/react/outline";
 
 /**
  * интерфейс кнопки добавления трека в плейлист
@@ -12,7 +14,7 @@ import { addToPlaylistButtonClicked } from "./model";
 interface AddToPlaylistButtonProps {
   track: Song;
 }
-export const AddToPlaylistButton = ({ track }: AddToPlaylistButtonProps) => {
+export const AddToPlaylistButton = memo<AddToPlaylistButtonProps>(({ track }) => {
   const handleButtonClicked = useUnit(addToPlaylistButtonClicked);
 
   const onClick = () => handleButtonClicked(track);
@@ -22,4 +24,6 @@ export const AddToPlaylistButton = ({ track }: AddToPlaylistButtonProps) => {
       <PlusSmIcon className="h-6 w-6 text-gray-500 duration-200 hover:animate-cross-spin hover:text-gray-900" />
     </button>
   );
-};
+});
+
+AddToPlaylistButton.displayName = "AddToPlaylistButton";
