+++ src/widgets/tracklist/index.ts
@@ -1,0 +1,1 @@
+export * from "./view";
